pyproject.toml
[project]
name = "email-monitor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imap-tools>=1.11.1",
    "imgkit>=1.2.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
